+++ src/main/java/com/vuducminh/stylash/service/UserService.java
@@ -1,14 +1,15 @@
 package com.vuducminh.stylash.service;
 
-import com.vuducminh.stylash.model.User;
+import com.vuducminh.stylash.user.User;
 
 import java.util.List;
 
 public interface UserService {
     List<User> getAllUsers();
-    User getUserById(Long userId);
+    User getUserById(Integer userId);
+
     User createUser(User user);
     User updateUser(User user);
-    void deleteUser(Long userId);
+    void deleteUser(Integer userId);
 }
